requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.2
django-cors-headers==3.14.0
django-formtools==2.4
django-ratelimit==3.0.1
djangorestframework==3.14.0
dnspython==2.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
idna==3.4
kombu==5.2.4
numpy==1.24.2
opencv-contrib-python==4.7.0.72
openpyxl==3.1.2
packaging==23.0
Pillow==9.2.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytesseract==0.3.10
python-dateutil==2.8.2
python-memcached==1.59
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
vine==5.0.0
wcwidth==0.2.6
